\name{cleanRprof}
\alias{cleanRprof}
\title{Clean Rprof}
\usage{
  cleanRprof(filename, namespace = FALSE)
}
\arguments{
  \item{filename}{A character string that contains either a
  relative or direct path to the file containing the Rprof
  output.}

  \item{namespace}{A logical value indicating if you want
  the namespace detected.}
}
\value{
  A nice data frame
}
\description{
  Takes a file corresponding to output from Rprof and
  cleans it up and produces a nice data frame.
}
\details{
  Does what I said
}
\author{
  Dason Kurkiewicz
}

